Add compatibility table to README
The table summarizes the various compatibility reports from the TobleMiner/HPE-FlexibleLOM-adapter issue tracker. It lists the card, adapter revision, and reference link(s) for each compatibility report. Modifications are noted and hyperlinked as mentioned in the issue threads.

The intro to the table also points out the existence of adapter revisions, and the community-sourced nature of compatibility reports.

Hat tip to @KCORES and their https://github.com/KCORES/KCORES-FlexibleLOM-Adapter with an even more featureful compat table